refactor: the mixnet surface of LightClient lives in lightclient/mixnet.rs
The nym-gated impl block (the slot lifecycle, the enables, attach,
disable, the status accessors, the route resolver, the probe), the
mixnet price fetch, and the two nym-gated contract-test modules move
from lightclient.rs into the new lightclient/mixnet.rs, declared as one
nym-gated module. Method resolution rides the type, so no call site or
import changes anywhere; vacate_mixnet_slot and publish_mixnet_slot_state
widen to pub(super) for their one caller (go_offline), matching their
previous effective visibility. What stays behind is what cannot move:
the gated struct fields, the constructor initializers, the gated hunks
inside ungated functions, and consent_to_clearnet_for_tests, whose
deliberate lack of a nym gate is its documented contract. The moved item
doc-comments are swept to one sentence per the 2026-08-10 documentation
rule, and the impl header and the two contract-mod docs become module
headers.
